fix memory leak in soundfile
Do not create an event listener every time a soundfile is played. Reference the same event listener and bind to this instance of the soundfile. We use this event listener to keep track of the current time (and may want to consider deprecating that functionality because it creates complications like this!).

Same for the "clearOnEnd" event listener, which we use to remove references to the bufferSourceNodes within a single sound file when they are done playing so that they can be garbage collected. This is useful to be able to control many bufferSourceNodes from a single sound file, and do things like soundfile.stopAll(), but it adds complexity and could also be a candidate for deprecation.